Factor out debug info producer helper methods
style fixes
